python_client_example/requirements.txt
python-socketio==5.0.4
pyttsx3==2.90
